crafted test and deploy script
test proves that tables are consistent.

deploy script does that too.

haven't tried to deploy when a gid is known missing, however, as I think
the geoms are all there in the real db.  Need to try messing with stuff
on a machine.  Or just run it on the broken machine, I guess.